dot1x(windows): harden syscall error paths; move WLAN XML parsing to shared file
Address Copilot review on the Windows backend:

- GetStatus: bounds-check dataSize against sizeof(wlanConnectionAttributes)
  before dereferencing the WlanQueryInterface buffer, returning a
  per-interface error on a short buffer instead of risking an OOB read.
- getWlanProfileXML / enumerateWlanInterfaceInfos: split the ret != 0 and
  nil-pointer cases so a "succeeded but returned no data" result no longer
  surfaces a misleading "errno 0" message.
- Move the pure-Go WLAN profile XML parsing (eapMethod/EAPType/authMode/
  TrustedRootCA extractors, isHexString, formatSHA1Hex, xml token helpers)
  and their tests out of the //go:build windows files into build-tag-free
  dot1x_wlanprofile.go / _test.go, so the parsing logic compiles, runs, and
  is coverage-counted on every platform (it now executes on the Linux CI /
  darwin test runs, not just Windows). Registered both files in BUILD.bazel.
